dal and domain emp

+++ dev/Assignment1/src/DomainLayer/DriverDL.java
@@ -9,7 +9,7 @@ public class DriverDL extends ShiftEmployee {
     public boolean isBusy = false;
 
     public DriverDL(int id, String name, LocationDL branch, String bankAccount, int salary, LocalDate startDate,
-            int vacationDays, int sickDays, double educationFund, double socialBenefits,
+            int vacationDays, int sickDays, float educationFund, float socialBenefits,
             String password, List<String> licenseType) {
         super(id, name, branch, bankAccount, salary, startDate,
                 vacationDays, sickDays, educationFund, socialBenefits, password, Role.DRIVER);
